+++ test/testdeadcall.lua
@@ -14,12 +14,10 @@ end)
 else
 
 	skynet.start(function()
-		local test = skynet.newservice("testdeadcall", "test")	-- launch self in test mode
+		local test = skynet.newservice(SERVICE_NAME, "test")	-- launch self in test mode
 
 		print(pcall(function()
-			skynet.send(test,"lua", "hello world")
-			skynet.send(test,"lua", "never get there")
-			skynet.call(test,"lua", "fake call")
+			skynet.call(test,"lua", "dead call")
 		end))
 
 		skynet.exit()