+++ HISTORY.md
@@ -1,6 +1,35 @@
+v0.4.2 (2014-7-14)
+-----------
+* Bugfix : invalid negative socket id 
+* Add optional TCP_NODELAY support
+* Add worker thread weight
+* Add skynet.queue
+* Bugfix: socketchannel
+* cluster can throw error
+* Add readline and writeline to clientsocket lib
+* Add cluster.reload to reload config file
+* Add datacenter.wait
+
+v0.4.1 (2014-7-7)
+-----------
+* Add SERVICE_NAME in loader
+* Throw error back when skynet.error
+* Add skynet.task
+* Bugfix for last version (harbor service bugs)
+
+v0.4.0 (2014-6-30)
+-----------
+* Optimize redis driver `compose_message`.
+* Add module skynet.harbor for monitor harbor connect/disconnect, see test/testharborlink.lua .
+* cluster.open support cluster name.
+* Add new api skynet.packstring , and skynet.unpack support lua string
+* socket.listen support put port into address. (address:port)
+* Redesign harbor/master/dummy, remove lots of C code and rewite in lua.
+* Remove block connect api, queue sending message during connecting now.
+* Add skynet.time()
+
 v0.3.2 (2014-6-23)
 ----------
-* Bugifx : mongo driver and lua-bson . (objectid encoding, and gc problem when mongdo driver reply bson object).
 * Bugfix : cluster (double free).
 * Add socket.header() to decode big-endian package header (and fix the bug in cluster).